@@ -1,11 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-config',
-  standalone: true,
-  imports: [],
-  templateUrl: './config.component.html',
-  styleUrl: './config.component.css'
+    selector: 'app-config',
+    imports: [],
+    templateUrl: './config.component.html',
+    styleUrl: './config.component.css'
 })
 export class ConfigComponent {
 
